[feat] #17 리뷰 반영해서 수정
