+++ src/main/java/com/miruni/backend/domain/user/controller/AuthController.java
@@ -1,8 +1,15 @@
 package com.miruni.backend.domain.user.controller;
 
+import com.miruni.backend.domain.user.dto.request.EmailVerificationRequest;
+import com.miruni.backend.domain.user.dto.request.EmailVerificationVerifyRequest;
 import com.miruni.backend.domain.user.dto.request.LoginRequest;
+import com.miruni.backend.domain.user.dto.request.ResetPasswordRequest;
+import com.miruni.backend.domain.user.dto.request.UserSignupRequest;
 import com.miruni.backend.domain.user.dto.response.JwtResponseDto;
+import com.miruni.backend.domain.user.dto.response.VerifyResponse;
 import com.miruni.backend.domain.user.service.AuthCommandService;
+import com.miruni.backend.domain.user.service.UserCommandService;
+import com.miruni.backend.domain.user.service.VerificationService;
 import com.miruni.backend.global.authroize.AuthToken;
 import com.miruni.backend.global.authroize.LoginUser;
 import jakarta.validation.Valid;
@@ -18,6 +25,8 @@ import lombok.RequiredArgsConstructor;
 public class AuthController implements AuthApi {
 
     private final AuthCommandService authCommandService;
+    private final UserCommandService userCommandService;
+    private final VerificationService verificationService;
 
     // 일반 로그인 API
     @PostMapping("/token")
@@ -29,6 +38,42 @@ public class AuthController implements AuthApi {
     @DeleteMapping("/token")
     public void logout(@AuthToken String accessToken, @LoginUser Long userId) {
         authCommandService.logout(accessToken, userId);
+    }
+
+    // 일반 회원가입 API
+    @PostMapping("/signup")
+    public JwtResponseDto signup(@Valid @RequestBody UserSignupRequest request) {
+        return userCommandService.signup(request);
+    }
+
+    // 회원가입 이메일 인증코드 요청
+    @PostMapping("/signup/email-verification")
+    public void requestEmailVerification(@Valid @RequestBody EmailVerificationRequest request) {
+        verificationService.sendSignUpVerificationCode(request.email());
+    }
+
+    // 회원가입 이메일 인증코드 검증
+    @PostMapping("/signup/email-verification/confirm")
+    public void verifyEmailVerification(@Valid @RequestBody EmailVerificationVerifyRequest request) {
+        verificationService.verifySignUpVerificationCode(request.email(), request.code());
+    }
+
+    // 비밀번호 재설정 요청
+    @PostMapping("/password/reset")
+    public void requestPasswordReset(@Valid @RequestBody EmailVerificationRequest request) {
+        verificationService.requestPasswordReset(request.email());
+    }
+
+    // 비밀번호 재설정 코드 검증 - resetToken 발급
+    @PostMapping("/password/reset/verification")
+    public VerifyResponse verifyPasswordResetCode(@Valid @RequestBody EmailVerificationVerifyRequest request) {
+        return verificationService.verifyPasswordResetCode(request);
+    }
+
+    // 비밀번호 재설정 완료 - resetToken으로 새 비밀번호 설정
+    @PostMapping("/password/reset/confirm")
+    public void resetPassword(@Valid @RequestBody ResetPasswordRequest request) {
+        userCommandService.resetPassword(request);
     }
 
 }
