+++ src/main/java/com/miruni/backend/domain/user/validator/UserValidator.java
@@ -1,6 +1,5 @@
 package com.miruni.backend.domain.user.validator;
 
-import com.miruni.backend.domain.user.dto.request.UserSignupRequest;
 import com.miruni.backend.domain.user.exception.UserErrorCode;
 import com.miruni.backend.global.exception.BaseException;
 import lombok.RequiredArgsConstructor;
@@ -11,11 +10,10 @@ import org.springframework.stereotype.Component;
 public class UserValidator {
 
     /**
-     * 약관 동의 검증
+     * 서비스 이용약관 동의 여부만 검증 (필수)
      */
-    public void validateAgreements(UserSignupRequest request) {
-        // 서비스 이용약관은 필수
-        if (request.serviceAgreed() == null || !request.serviceAgreed()) {
+    public void validateAgreements(Boolean serviceAgreed) {
+        if (serviceAgreed == null || !serviceAgreed) {
             throw BaseException.type(UserErrorCode.AGREEMENT_REQUIRED);
         }
     }
